Fix warnings related to using size_t typed index to access Kokkos::Array data

+++ src/kalypsso/core/CheckFaceBorderCompatibility.cpp
@@ -96,7 +96,7 @@ CheckFaceBorderCompatibility<dim, device_t>::check(index_t const & surface_flati
   // if (iOct_neigh >= m_amr_mesh_info.local_num_quadrants())
   //  return;
 
-  const auto face_dir = ijk_face[dim];
+  const auto face_dir = static_cast<size_t>(ijk_face[dim]);
 
   // get a face location at neighbor location
   // we only explore 2 cases:
